homework-server/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
pymysql==1.1.0
cryptography==42.0.4
alembic==1.13.1
redis==5.0.1
pyjwt==2.8.0
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
apscheduler==3.10.4
python-dotenv==1.0.1
python-multipart==0.0.9
